+++ example/android/app/src/main/java/com/dexterous/flutterlocalnotificationsexample/MainActivity.java
@@ -1,14 +1,36 @@
 package com.dexterous.flutterlocalnotificationsexample;
 
+import android.content.ContentResolver;
+import android.content.Context;
 import android.os.Bundle;
 
 import io.flutter.app.FlutterActivity;
+import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugins.GeneratedPluginRegistrant;
 
 public class MainActivity extends FlutterActivity {
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
-  }
+    private static String resourceToUriString(Context context, int resId) {
+        return
+                ContentResolver.SCHEME_ANDROID_RESOURCE
+                        + "://"
+                        + context.getResources().getResourcePackageName(resId)
+                        + "/"
+                        + context.getResources().getResourceTypeName(resId)
+                        + "/"
+                        + context.getResources().getResourceEntryName(resId);
+    }
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        GeneratedPluginRegistrant.registerWith(this);
+        new MethodChannel(getFlutterView(), "crossingthestreams.io/resourceResolver").setMethodCallHandler(
+                (call, result) -> {
+                    if ("drawableToUri".equals(call.method)) {
+                        int resourceId = getResources().getIdentifier((String) call.arguments, "drawable", getPackageName());
+                        String uriString = resourceToUriString(getApplicationContext(), resourceId);
+                        result.success(uriString);
+                    }
+                });
+    }
 }