+++ lib/src/platform_specifics/android/person.dart
@@ -1,18 +1,32 @@
 part of flutter_local_notifications;
 
+/// Details of a person e.g. someone who sent a message
 class Person {
+  /// Whether or not this person represents a machine rather than a human.
   final bool bot;
+
+  /// Icon for this person
   final String icon;
-  final BitmapSource iconBitmapSource;
+
+  /// Determines how the icon should be interpreted/resolved e.g. as a drawable
+  final IconSource iconSource;
+
+  /// Whether or not this is an important person
   final bool important;
+
+  /// Unique identifier for this person
   final String key;
+
+  /// Name of this person
   final String name;
+
+  /// Uri for this person
   final String uri;
 
   Person(
       {this.bot,
       this.icon,
-      this.iconBitmapSource,
+      this.iconSource,
       this.important,
       this.key,
       this.name,
@@ -22,7 +36,7 @@ class Person {
     return <String, dynamic>{
       'bot': bot,
       'icon': icon,
-      'iconBitmapSource': iconBitmapSource?.index,
+      'iconSource': iconSource?.index,
       'important': important,
       'key': key,
       'name': name,