+++ lib/src/platform_specifics/android/styles/messaging_style_information.dart
@@ -1,9 +1,17 @@
 part of flutter_local_notifications;
 
+/// Used to pass the content for an Android notification displayed using the messaging style
 class MessagingStyleInformation extends DefaultStyleInformation {
+  /// The person displayed for any messages that are sent by the user
   final Person person;
+
+  /// The title to be displayed on this conversation
   final String conversationTitle;
+
+  /// Whether this conversation notification represents a group
   final bool groupConversation;
+
+  /// Messages to be displayed by this notification
   final List<Message> messages;
   MessagingStyleInformation(this.person,
       {this.conversationTitle,