+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# [0.5.1]
+* Updated gradle plugin of example app to 3.3.1
+* Added support for messaging style notifications on Android as requested in issue [159](https://github.com/MaikuB/flutter_local_notifications/issues/159). See example app for some sample code
+
 # [0.5.0]
 * **BREAKING CHANGE** Migrated to use AndroidX as the Android support libraries are deprecated. There shouldn't be any functional changes. Developers may require migrating their apps to support this following [this guide](https://developer.android.com/jetpack/androidx/migrate). This addresses issue [162]. (https://github.com/MaikuB/flutter_local_notifications/issues/162). Thanks to Jeff Scaturro (JeffScaturro) for submitting the PR for this work. Note that if you don't want to migrate your app to use AndroidX yet then you may need to pin dependencies to a specific version